+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+* Nicks (or usernames) are now saved when a member is updated
+* Join and part messages now show nickname or previous name where applicable
 
 
 ## [1.0.1] - 2020-06-03