+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [1.1.0] - 2020-06-04
 ### Added
 * Nicks (or usernames) are now saved when a member is updated
 * Join and part messages now show nickname or previous name where applicable
@@ -29,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Global permission checks - messages from members without Administrator are ignored
 
 
-[Unreleased]: https://github.com/dshoreman/smegbot/compare/v1.0.1...develop
+[Unreleased]: https://github.com/dshoreman/smegbot/compare/v1.1.0...develop
+[1.1.0]: https://github.com/dshoreman/smegbot/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/dshoreman/smegbot/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/dshoreman/smegbot/releases/tag/v1.0.0